Factor out image.rs
